requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
langchain-experimental==0.3.4
langgraph==0.4.0
openai==1.75.0
duckduckgo-search==7.5.3
python-dotenv==1.1.0
